Added classes and packages
Added some packages.

Using a function in Visual Paradigm, I then generated Java classes based on "UML Diagram 1.0.vpp". I had to make some temporary changes to the diagram in order to use the function, in this case removing the {id} part of the name and the multiplicity of some variables.

I also manually added the package info to the classes in my IDE.